requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.100.0
python-dotenv==1.0.0

